try/catch block and async
This command sometimes falsely returns a did not respond error, I presume because the uploaded image function takes too long. So I've set the command to be async and wait on reading the image data.

I've also put the whole thing in a try catch statement so that if it breaks, it can just output an error instead of imploding the bot. That's a good thing, right?