slash command params
Moved the "random" parameter to appear first since it will likely be used more often. This parameter is now a string instead of a boolean so that a pre-determined choice can be shown and it'll only auto-suggest TRUE since there was no point in having both true and false from the boolean.