fixed toggling, button IDs, try/catch
Previously all buttons created with this were given button1 as an ID and I think that broke interactions. Now they're given an ID specific to the role given to them, which I think works fine?

Also the previous code didn't properly recognize whether or not the interacting user had the role, so toggling off didn't work. It should now.

Now it's all in a try catch block, because that seems prudent. Sometimes when this command broke, it would force the bot to restart. Shouldn't have to now, I think.